refactor: onSubmit moved to its own hook
